In this assignment, you will be implementing a LRU (least-recently-used) cache which keeps most recently accessed items in the cache using LRU strategy. In the LRU strategy when a new item needs to be stored in cache with no more capacity left in cache, the item that had been accessed earliest in the past needs to be evicted to make room for the new item. LRU cache has a Persister instance (provided in the constructor) to store new or modified items when they are added/modified. Also when an accessed item is not in cache it is retrieved using Persister instance and then stored in cache. This is a write-through cache as opposed to write-back cache when a modified item is written to persistent store in a lazy manner.
You should complete the generic class LRUCache that had been provided to you. Do not change the interfaces Cacheable and Persister. Complete the constructor and other public methods in the class along with CacheIterator implementation methods. Read the Javadoc comments to understand what needs to be implemented in the methods.

I have also provided two instance variables itemMap and lruList to help you with the implementations. You can add more variables if necessary. You need to instantiate these variables in the constructor and use them as appropriate in various methods. The method getFaultRate() gives percent of accesses (using getItem() and putItem() for an existing key) that requires the Persister instance to retrieve or modify the item. Keep appropriate counters to calculate this value. The method resetFaultRateStats() resets these counters so that only accesses from that point onwards should be calculated for the next getFaultRate() call.

For your testing purposes, I provided a sample Persister implementation (SimpleFakePersister) and a sample Cacheable implementation (SimpleCacheItem). I may use a different implementation of these interfaces for my testing and your LRUCache should work for these.